Move all packages from /pkg to /internal
