+++ internal/resolve/catalog_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/operator-framework/operator-registry/alpha/property"
 
 	ocv1alpha1 "github.com/operator-framework/operator-controller/api/v1alpha1"
-	"github.com/operator-framework/operator-controller/pkg/features"
+	"github.com/operator-framework/operator-controller/internal/features"
 )
 
 func TestInvalidClusterExtensionVersionRange(t *testing.T) {
