+++ test/upgrade-e2e/upgrade_e2e_suite_test.go
@@ -9,7 +9,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/operator-framework/operator-controller/pkg/scheme"
+	"github.com/operator-framework/operator-controller/internal/scheme"
 )
 
 const (
